add demangler tests for testing label demangling
