+++ test/demangle-cases/bug-1468b.asm.json
@@ -1,0 +1,37 @@
+{
+  "asm": [
+    {
+      "labels": [],
+      "source": null,
+      "text": "g():",
+    },
+    {
+      "labels": [],
+      "source": null,
+      "text": "        subq    $8, %rsp",
+    },
+    {
+      "labels": [],
+      "source": null,
+      "text": "        call    *foo::baz+8(%rip)",
+    },
+    {
+      "labels": [],
+      "source": null,
+      "text": "        xorl    %eax, %eax",
+    },
+    {
+      "labels": [],
+      "source": null,
+      "text": "        addq    $8, %rsp",
+    },
+    {
+      "labels": [],
+      "source": null,
+      "text": "        ret",
+    },
+  ],
+  "labelDefinitions": {
+    "g()": 1,
+  },
+}
