+++ test/demangle-cases/similar-symbols.asm.json
@@ -1,0 +1,155 @@
+{
+  "asm": [
+    {
+      "labels": [],
+      "source": null,
+      "text": "aa(int):",
+    },
+    {
+      "labels": [],
+      "source": null,
+      "text": "        push    rbp",
+    },
+    {
+      "labels": [],
+      "source": null,
+      "text": "        mov     rbp, rsp",
+    },
+    {
+      "labels": [],
+      "source": null,
+      "text": "        mov     DWORD PTR [rbp-4], edi",
+    },
+    {
+      "labels": [],
+      "source": null,
+      "text": "        nop",
+    },
+    {
+      "labels": [],
+      "source": null,
+      "text": "        pop     rbp",
+    },
+    {
+      "labels": [],
+      "source": null,
+      "text": "        ret",
+    },
+    {
+      "labels": [],
+      "source": null,
+      "text": "aa(int, int):",
+    },
+    {
+      "labels": [],
+      "source": null,
+      "text": "        push    rbp",
+    },
+    {
+      "labels": [],
+      "source": null,
+      "text": "        mov     rbp, rsp",
+    },
+    {
+      "labels": [],
+      "source": null,
+      "text": "        mov     DWORD PTR [rbp-4], edi",
+    },
+    {
+      "labels": [],
+      "source": null,
+      "text": "        mov     DWORD PTR [rbp-8], esi",
+    },
+    {
+      "labels": [],
+      "source": null,
+      "text": "        nop",
+    },
+    {
+      "labels": [],
+      "source": null,
+      "text": "        pop     rbp",
+    },
+    {
+      "labels": [],
+      "source": null,
+      "text": "        ret",
+    },
+    {
+      "labels": [],
+      "source": null,
+      "text": "main:",
+    },
+    {
+      "labels": [],
+      "source": null,
+      "text": "        push    rbp",
+    },
+    {
+      "labels": [],
+      "source": null,
+      "text": "        mov     rbp, rsp",
+    },
+    {
+      "labels": [],
+      "source": null,
+      "text": "        mov     esi, 1",
+    },
+    {
+      "labels": [],
+      "source": null,
+      "text": "        mov     edi, 0",
+    },
+    {
+      "labels": [
+        {
+          "name": "aa(int, int)",
+          "range": {
+            "endCol": 29,
+            "startCol": 17,
+          },
+        },
+      ],
+      "source": null,
+      "text": "        call    aa(int, int)",
+    },
+    {
+      "labels": [],
+      "source": null,
+      "text": "        mov     edi, 1",
+    },
+    {
+      "labels": [
+        {
+          "name": "aa(int)",
+          "range": {
+            "endCol": 24,
+            "startCol": 17,
+          },
+        },
+      ],
+      "source": null,
+      "text": "        call    aa(int)",
+    },
+    {
+      "labels": [],
+      "source": null,
+      "text": "        mov     eax, 0",
+    },
+    {
+      "labels": [],
+      "source": null,
+      "text": "        pop     rbp",
+    },
+    {
+      "labels": [],
+      "source": null,
+      "text": "        ret",
+    },
+  ],
+  "labelDefinitions": {
+    "aa(int)": 1,
+    "aa(int, int)": 8,
+    "main": 16,
+  },
+}
